Cleaning up react Meetup API Endpoint
